Sort of fixes weird settings sizing, fixes #73 weird spacing that
appeared sometimes in the flexbox'ed elements. The issue was
padding which was set as a percent. Child padding values based on
percent didn't seem to mix well with flexbox and rendered child
elements too small sometimes. Go figure. Pixel values work much
better.

Resetting sizes is only based on a setTimeout delay, which won't
perform consistently on environments with too much lag. TODO:
find a way of detecting the right moment to reset the size.