Changes described below for getting the punctuation out of the fragment before finding the indicator character position. Positions at 2nd character if there are less than 7, 3rd character if 7-9, and 4th character if greater.
Just adding classes for the spans

Using the length of the word with only letters.

Deals with shifting the focus character if we land on something that's not a letter.

Using a loop so can deal with any number of punctuation symbols encountered in the "middle" text that shows under the indicator.

Skip shifting characters if we only have a 1 letter word.

Updated regex to look for particular symbols to exclude.

Fixes for symbols in the first part of the word, shifting things over x number of spaces where x is number of non character symbols encountered.

Greater than or EQUAL TO

Fixing off by one error for the indicator position.

Just a little bit of tidying.

Little more cleanup and fixed the > on 10 to be >= since only 7-9 goes to position 3